refactor: página de clientes e de carrinho
