+++ src/main/webapp/css/paginas/admin/clientes.css
@@ -1,9 +1,7 @@
-@import url(/css/style.css);
 @import url(/css/crud_cliente.css);
 
 main {
     display: grid;
-    width: 90%;
 }
 
 #header-clientes, #tabela-clientes {
