+++ src/main/webapp/css/paginas/carrinho.css
@@ -1,5 +1,3 @@
-@import url(../../../../css/style.css);
-
 .celula-produto {
     justify-items: center;
     padding: 8px;
