+++ src/main/webapp/paginas/carrinho.html
@@ -1,0 +1,36 @@
+<title>Carrinho</title>
+
+<nav id="main-nav"></nav>
+
+<div id="main-container" class="container my-5 flex-grow-1">
+  <h1 class="mb-4 text-center d-flex align-items-center justify-content-center gap-2">
+    <img src="/img/carrinho.png">
+    Carrinho
+  </h1>
+
+  <!-- Cart Table -->
+  <div class="table-responsive">
+    <table class="table table-bordered table-hover" id="cartTable">
+      <thead class="table-dark">
+        <tr>
+          <th>Produto</th>
+          <th>Preço</th>
+          <th>Quantidade</th>
+          <th>Total</th>
+          <th>Operações</th>
+        </tr>
+      </thead>
+      <tbody id="tabela-produtos-carrinho"></tbody>
+    </table>
+  </div>
+
+  <!-- Cart Summary -->
+  <div class="d-flex justify-content-between align-items-center mt-4">
+    <h4>
+      Total: <span class="text-success" id="total-carrinho"></span>
+    </h4>
+    <a href="/compra" id="link-compra"><button class="btn btn-primary btn-lg">Comprar</button></a>
+  </div>
+</div>
+
+<footer id="main-footer"></footer>
