--- src/main/webapp/paginas/cliente/carrinho/carrinho.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Carrinho</title>
-  <link rel="stylesheet" href="../../../paginas/cliente/carrinho/css/carrinho.css">
-</head>
-
-<body class="bg-light d-flex flex-column">
-
-  <nav id="main-nav"></nav>
-
-  <div id="main-container" class="container my-5 flex-grow-1">
-    <h1 class="mb-4 text-center d-flex align-items-center justify-content-center gap-2">
-      <img src="/img/carrinho.png">
-      Carrinho
-    </h1>
-
-    <!-- Cart Table -->
-    <div class="table-responsive">
-      <table class="table table-bordered table-hover" id="cartTable">
-        <thead class="table-dark">
-          <tr>
-            <th>Produto</th>
-            <th>Preço</th>
-            <th>Quantidade</th>
-            <th>Total</th>
-            <th>Operações</th>
-          </tr>
-        </thead>
-        <tbody id="tabela-produtos-carrinho"></tbody>
-      </table>
-    </div>
-
-    <!-- Cart Summary -->
-    <div class="d-flex justify-content-between align-items-center mt-4">
-      <h4>
-        Total: <span class="text-success" id="total-carrinho"></span>
-      </h4>
-      <a href="/compra" id="link-compra"><button class="btn btn-primary btn-lg">Comprar</button></a>
-    </div>
-  </div>
-
-  <footer id="main-footer"></footer>
-
-  <script type="module" src="/js/loadPage.js"></script>
-  <script type="module" src="/paginas/cliente/carrinho/js/carrinho.js"></script>
-</body>
-
-</html>
